+++ include/openmc/nuclide.h
@@ -109,13 +109,13 @@ private:
 struct Nuclide_ {
   struct EnergyGrid_ {
     rtBufferId<int, 1> grid_index;
-    rtBufferId<double, 1> energy;
+    rtBufferId<float, 1> energy;
 
     __forceinline__ __device__ EnergyGrid_() {}
 
     __forceinline__ __device__ EnergyGrid_(Nuclide::EnergyGrid &e,
                                            rtBufferId<int, 1> grid_index,
-                                           rtBufferId<double, 1> energy) {
+                                           rtBufferId<float, 1> energy) {
       this->grid_index = grid_index;
       this->energy = energy;
     }
@@ -126,13 +126,13 @@ struct Nuclide_ {
   int Z_; //!< Atomic number
   int A_; //!< Mass number
   int metastable_; //!< Metastable state
-  double awr_; //!< Atomic weight ratio
+  float awr_; //!< Atomic weight ratio
   int i_nuclide_; //!< Index in the nuclides array
 
   // Temperature dependent cross section data
-  rtBufferId<double, 1> kTs_; //!< temperatures in eV (k*T)
+  rtBufferId<float, 1> kTs_; //!< temperatures in eV (k*T)
   rtBufferId<EnergyGrid_, 1> grid_; //!< Energy grid at each temperature
-  rtBufferId<rtBufferId<double, 1>, 1> xs_; //!< Cross sections at each temperature
+  rtBufferId<rtBufferId<float, 1>, 1> xs_; //!< Cross sections at each temperature
 
   // Fission data
   bool fissionable_; //!< Whether nuclide is fissionable
@@ -146,9 +146,9 @@ struct Nuclide_ {
   __forceinline__ __device__ Nuclide_() {}
 
   __forceinline__ __device__ Nuclide_(Nuclide *n,
-                                      rtBufferId<double, 1> kTs_,
+                                      rtBufferId<float, 1> kTs_,
                                       rtBufferId<EnergyGrid_> grid_,
-                                      rtBufferId<rtBufferId<double, 1>, 1> xs_,
+                                      rtBufferId<rtBufferId<float, 1>, 1> xs_,
                                       rtBufferId<Reaction_, 1> fission_rx_,
                                       rtBufferId<Reaction_, 1> reactions_,
                                       rtBufferId<int, 1> index_inelastic_scatter) {
